Fix Plasma (non-Qt) icon scaling with integer scale factors
An embarrassing error in 4edb830f53aac4833ece024df59441029b01fde6
caused the function which returns an appropriate scale factor for icons
to both take and return integer values rather than floating point
values, preventing it from working properly with any scale factor other
than an integer scale factor (e.g. 200%). This commit fixes that issue.

+++ src/declarativeimports/core/units.cpp
@@ -174,7 +174,7 @@ int Units::roundToIconSize(int size)
     }
 }
 
-int Units::bestIconScaleForDevicePixelRatio(const int ratio)
+qreal Units::bestIconScaleForDevicePixelRatio(const qreal ratio)
 {
     if (ratio < 1.5) {
         return 1;
